+++ pkg/apis/core/v1alpha1/generated.proto
@@ -394,6 +394,7 @@ message ContainerRuntime {
 message ControlPlane {
   // HighAvailability holds the configuration settings for high availability of the
   // control plane of a shoot.
+  // +optional
   optional HighAvailability highAvailability = 1;
 }
